Add abstract JDBC store
