+++ jdbc-adapter/src/main/java/ch/unibas/dmi/dbis/polyphenydb/adapter/jdbc/stores/HsqldbStore.java
@@ -3,19 +3,11 @@ package ch.unibas.dmi.dbis.polyphenydb.adapter.jdbc.stores;
 
 import ch.unibas.dmi.dbis.polyphenydb.PolySqlType;
 import ch.unibas.dmi.dbis.polyphenydb.PolyXid;
-import ch.unibas.dmi.dbis.polyphenydb.Store;
 import ch.unibas.dmi.dbis.polyphenydb.Transaction;
 import ch.unibas.dmi.dbis.polyphenydb.adapter.jdbc.JdbcPhysicalNameProvider;
 import ch.unibas.dmi.dbis.polyphenydb.adapter.jdbc.JdbcSchema;
 import ch.unibas.dmi.dbis.polyphenydb.catalog.entity.CatalogColumn;
 import ch.unibas.dmi.dbis.polyphenydb.catalog.entity.combined.CatalogCombinedTable;
-import ch.unibas.dmi.dbis.polyphenydb.information.InformationGraph;
-import ch.unibas.dmi.dbis.polyphenydb.information.InformationGraph.GraphData;
-import ch.unibas.dmi.dbis.polyphenydb.information.InformationGraph.GraphType;
-import ch.unibas.dmi.dbis.polyphenydb.information.InformationGroup;
-import ch.unibas.dmi.dbis.polyphenydb.information.InformationManager;
-import ch.unibas.dmi.dbis.polyphenydb.information.InformationPage;
-import ch.unibas.dmi.dbis.polyphenydb.information.InformationTable;
 import ch.unibas.dmi.dbis.polyphenydb.jdbc.Context;
 import ch.unibas.dmi.dbis.polyphenydb.schema.Schema;
 import ch.unibas.dmi.dbis.polyphenydb.schema.SchemaPlus;
@@ -27,19 +19,15 @@ import java.io.File;
 import java.sql.Connection;
 import java.sql.SQLException;
 import java.sql.Statement;
-import java.util.Arrays;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
-import java.util.concurrent.Executors;
-import java.util.concurrent.ScheduledExecutorService;
-import java.util.concurrent.TimeUnit;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.dbcp2.BasicDataSource;
 
 
 @Slf4j
-public class HsqldbStore extends Store {
+public class HsqldbStore extends AbstractJdbcStore {
 
     @SuppressWarnings("WeakerAccess")
     public static final String ADAPTER_NAME = "HSQLDB";
@@ -75,31 +63,8 @@ public class HsqldbStore extends Store {
         this.dataSource = dataSource;
         dialect = JdbcSchema.createDialect( SqlDialectFactoryImpl.INSTANCE, dataSource );
 
-        // ----------- Information Manager -----------
-        final InformationPage informationPage = new InformationPage( uniqueName, uniqueName );
-        final InformationGroup informationGroupConnectionPool = new InformationGroup( "JDBC Connection Pool", informationPage.getId() );
-
-        InformationManager im = InformationManager.getInstance();
-        im.addPage( informationPage );
-        im.addGroup( informationGroupConnectionPool );
-
-        InformationGraph connectionPoolSizeGraph = new InformationGraph(
-                "connectionPoolSizeGraph",
-                informationGroupConnectionPool.getId(),
-                GraphType.DOUGHNUT,
-                new String[]{ "Active", "Available", "Idle" }
-        );
-        im.registerInformation( connectionPoolSizeGraph );
-
-        InformationTable connectionPoolSizeTable = new InformationTable(
-                "connectionPoolSizeTable",
-                informationGroupConnectionPool.getId(),
-                Arrays.asList( "Attribute", "Value" ) );
-        im.registerInformation( connectionPoolSizeTable );
-
-        ConnectionPoolSizeInfo connectionPoolSizeInfo = new ConnectionPoolSizeInfo( connectionPoolSizeGraph, connectionPoolSizeTable );
-        ScheduledExecutorService exec = Executors.newSingleThreadScheduledExecutor();
-        exec.scheduleAtFixedRate( connectionPoolSizeInfo, 0, 5, TimeUnit.SECONDS );
+        // Register the JDBC Pool Size as information in the information manager
+        registerJdbcPoolSizeInformation( uniqueName, dataSource );
     }
 
 
@@ -277,6 +242,7 @@ public class HsqldbStore extends Store {
     @Override
     public void shutdown() {
         try {
+            removeInformationPage();
             dataSource.getConnection().createStatement().execute( "SHUTDOWN" );
             dataSource.close();
         } catch ( SQLException e ) {
@@ -340,38 +306,6 @@ public class HsqldbStore extends Store {
                     log.info( "Exception while closing connection!", e );
                 }
             }
-        }
-    }
-
-
-    private class ConnectionPoolSizeInfo implements Runnable {
-
-        private final InformationGraph graph;
-        private final InformationTable table;
-
-
-        ConnectionPoolSizeInfo( InformationGraph graph, InformationTable table ) {
-            this.graph = graph;
-            this.table = table;
-        }
-
-
-        @Override
-        public void run() {
-            int idle = dataSource.getNumIdle();
-            int active = dataSource.getNumActive();
-            int max = dataSource.getMaxTotal();
-            int available = max - idle - active;
-
-            graph.updateGraph(
-                    new String[]{ "Active", "Available", "Idle" },
-                    new GraphData<>( "hsqldb-connection-pool-data", new Integer[]{ active, available, idle } )
-            );
-
-            table.reset();
-            table.addRow( "Active", "" + active );
-            table.addRow( "Idle", "" + idle );
-            table.addRow( "Max", "" + max );
         }
     }
